requirements.txt
asgiref==3.5.0
astroid==2.3.3
certifi==2021.10.8
charset-normalizer==2.0.10
cloudinary==1.25.0
colorama==0.4.3
dj-database-url==0.5.0
Django==3.2
django-cloudinary-storage==0.3.0
django-crispy-forms==1.9.0
django-endless-pagination==2.0
django-heroku==0.3.1
django-hitcount==1.3.1
django-mdeditor==0.1.20
django-summernote==0.8.11.6
django-taggit==2.1.0
django-watson==1.5.4
gunicorn==20.0.4
idna==3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient
Pillow==9.0.0
psycopg2==2.9.3
pylint==2.4.4
python-decouple==3.1
pytz==2019.3
requests==2.27.1
six==1.14.0
sqlparse==0.3.1
urllib3==1.26.8
uWSGI==2.0.20
virtualenv==16.7.9
whitenoise==5.0.1
wrapt==1.11.2
